Combat finalized, menus finalized
Combat is now made and the menus are all working. All that is left programming wise is the upgrade system, and player UI. Enemy AI might also need to be improved to move around objects, but oh well. Once that is done, it's just animations, sprites, and making more rooms.